requirements.txt
alembic==1.8.1
amqp==5.1.1
astroid==2.11.7
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
billiard==4.2.0
blinker==1.6.2
boto3==1.24.33
botocore==1.27.33
celery==5.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coloredlogs==15.0.1
cryptography==41.0.2
Deprecated==1.2.13
dill==0.3.5.1
docutils==0.19
flasgger==0.9.7.1
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
GeoAlchemy2==0.14.1
gevent==24.2.1
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
gunicorn==20.1.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.3
intervals==0.9.2
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.7.2
kombu==5.3.7
lazy-object-proxy==1.7.1
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.7.0
mistune==3.0.1
moviepy==1.0.3
msgpack==1.0.4
neovim==0.3.1
numpy==2.0.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2==2.9.6
pycparser==2.21
pycountry==22.3.5
PyJWT==2.4.0
pylint==2.14.5
pynvim==0.4.3
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
pytz==2022.1
PyYAML==6.0.1
raven==6.10.0
redis==5.0.7
requests==2.28.1
s3transfer==0.6.0
sentry-sdk==2.12.0
shapely==2.0.4
sendgrid==6.11.0
six==1.16.0
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.3
tomli==2.0.1
tomlkit==0.11.1
twilio==7.11.0
urllib3==1.26.11
uuid==1.30
vine==5.1.0
voluptuous==0.13.1
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
Unidecode==1.3.8